# openapi/spec.yaml
openapi: 3.1.0
info:
  title: OAuth Implementation Server
  description: |
    A comprehensive OAuth 2.0 / OpenID Connect implementation server for testing all OAuth flows and standards.

    Supports: Authorization Code + PKCE, Client Credentials, Device Code (RFC 8628),
    CIBA (Client-Initiated Backchannel Authentication), PAR (RFC 9126), DPoP (RFC 9449),
    Token Introspection (RFC 7662), Token Revocation (RFC 7009), Dynamic Client Registration (RFC 7591),
    Resource Indicators (RFC 8707), OIDC Discovery, JWKS, UserInfo, TOTP-based MFA,
    Consent Management, Custom Scopes, and login_hint/prompt/max_age parameters.
  version: 1.1.0
  contact:
    name: API Support
    url: https://github.com/bravo68web/oauth-impl
servers:
  - url: https://localhost:8080
    description: Local development server
...
paths:
...
  /oauth/authorize:
    get:
      tags: [OAuth 2.0]
      summary: Authorization endpoint
      description: |
        Initiates the OAuth 2.0 authorization flow. The user-agent is redirected to this endpoint
        to authenticate the user and obtain consent. On success, redirects back to the client
        with an authorization code.

        For testing, pass `user_id` to auto-approve without user interaction.

        Supports PKCE via `code_challenge` and `code_challenge_method` parameters.
      operationId: authorize
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: The client identifier obtained from registration
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum:
              [
                code,
                token,
                id_token,
                "code id_token",
                "code token",
                "code id_token token",
                none,
              ]
          description: |
            OAuth response type:
            - `code`: Authorization code flow (query parameter)
            - `token`: Implicit access token (fragment)
            - `id_token`: Implicit ID token (fragment)
            - `code id_token`: Hybrid - code + ID token (fragment)
            - `code token`: Hybrid - code + access token (fragment)
            - `code id_token token`: Hybrid - all three (fragment)
            - `none`: No tokens returned
          example: code
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
          description: Must match one of the registered redirect URIs
          example: "https://example.com/callback"
        - name: scope
          in: query
          schema:
            type: string
          description: Space-separated list of requested scopes
          example: "openid profile email"
        - name: state
          in: query
          schema:
            type: string
          description: Opaque value for CSRF protection (recommended)
          example: "abc123xyz"
        - name: code_challenge
          in: query
          schema:
            type: string
          description: PKCE code challenge (base64url-encoded SHA256 of code_verifier)
          example: "E9Melhoa2OwvFrEMTJguCHaoeK1t8URWbuGJSstw-cM"
        - name: code_challenge_method
          in: query
          schema:
            type: string
            enum: [S256, plain]
          description: PKCE challenge method (S256 recommended)
          example: S256
        - name: nonce
          in: query
          schema:
            type: string
          description: OpenID Connect nonce for replay protection
          example: "n-0S6_WzA2Mj"
        - name: prompt
          in: query
          schema:
            type: string
            enum: [none, login, consent, select_account]
          description: |
            OIDC prompt parameter:
            - `none`: No UI shown; fail if not already authenticated and consented
            - `login`: Force re-authentication even if already logged in
            - `consent`: Force consent screen even if previously consented
            - `select_account`: Account chooser (not yet implemented)
          example: consent
        - name: login_hint
          in: query
          schema:
            type: string
          description: Hint to pre-fill the username on the login page
          example: "johndoe"
        - name: resource
          in: query
          schema:
            type: string
          description: |
            Resource indicator (RFC 8707). The URI of the resource server the client wants to access.
            The issued token will be scoped to this resource.
          example: "https://api.example.com"
        - name: user_id
          in: query
          schema:
            type: string
          description: User ID for auto-approval (testing only, not part of standard OAuth)
          example: "b5c6d7e8-f9a0-1234-5678-9abcdef01234"
      responses:
        "302":
          description: |
            Redirect to client with authorization code. The `Location` header contains the
            redirect URI with `code` and `state` query parameters.
          headers:
            Location:
              description: Redirect URI with authorization code
              schema:
                type: string
                example: "https://example.com/callback?code=SplxlOBeZQQYbYS6WxSbIA&state=abc123xyz"
        "400":
          description: Invalid request (missing parameters, invalid client, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
              example:
                error: invalid_request
                error_description: "client_id is required"
                state: "abc123xyz"
...
  /api/users:
    get:
      tags: [Management]
      security: [{ bearerAuth: [] }]
      summary: List all users
      description: Returns all registered users.
      operationId: listUsers
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
              example:
                - id: "b5c6d7e8-f9a0-1234-5678-9abcdef01234"
                  username: "johndoe"
                  email: "john@example.com"
                  phone_number: "+1-555-123-4567"
                  created_at: "2026-06-16T10:00:00Z"
    post:
      tags: [Management]
      security: [{ bearerAuth: [] }]
      summary: Create a new user
      description: Creates a new user account with hashed password storage.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            example:
              username: "johndoe"
              password: "secureP@ss123"
              email: "john@example.com"
              phone_number: "+1-555-123-4567"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              example:
                id: "b5c6d7e8-f9a0-1234-5678-9abcdef01234"
                username: "johndoe"
                email: "john@example.com"
                phone_number: "+1-555-123-4567"
                created_at: "2026-06-16T10:00:00Z"
        "400":
          description: Invalid request (username already exists, missing fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
              example:
                error: invalid_request
                error_description: "username already exists"
  /api/users/{userID}:
    get:
      tags: [Management]
      security: [{ bearerAuth: [] }]
      summary: Get user by ID
      description: Returns a single user by their ID.
      operationId: getUser
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: string
          description: The user ID
          example: "b5c6d7e8-f9a0-1234-5678-9abcdef01234"
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
    patch:
      tags: [Management]
      summary: Update a user profile or attributes
      description: Management can set `attributes`. Self-service `PATCH /api/me` cannot.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated user
        "404":
          description: User not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque OAuth 2.0 access token. Management routes require a client-credentials token whose scope contains management.
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic authentication with client_id:client_secret
  schemas:
...
    CreateUserRequest:
      type: object
      required: [username, password]
      description: Request body for creating a user
      example:
        username: "johndoe"
        password: "secureP@ss123"
        email: "john@example.com"
        phone_number: "+1-555-123-4567"
      properties:
        username:
          type: string
          description: Unique username
          example: "johndoe"
        password:
          type: string
          format: password
          description: User password (stored with the hasher in internal/hashalgo/algo.go)
          example: "secureP@ss123"
        email:
          type: string
          format: email
          description: User email address
          example: "john@example.com"
        phone_number:
          type: string
          description: User phone number (E.164 format recommended)
          example: "+1-555-123-4567"
    User:
      type: object
      description: User representation (password hash is never returned)
      example:
        id: "b5c6d7e8-f9a0-1234-5678-9abcdef01234"
        username: "johndoe"
        email: "john@example.com"
        phone_number: "+1-555-123-4567"
        created_at: "2026-06-16T10:00:00Z"
      properties:
        id:
          type: string
          description: Unique user identifier
          example: "b5c6d7e8-f9a0-1234-5678-9abcdef01234"
        username:
          type: string
          description: Username
          example: "johndoe"
        email:
          type: string
          description: Email address
          example: "john@example.com"
        phone_number:
          type: string
          description: Phone number
          example: "+1-555-123-4567"
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2026-06-16T10:00:00Z"

    # ── Token ──
...
    OAuthError:
      type: object
      required: [error]
      description: OAuth 2.0 error response
      example:
        error: invalid_request
        error_description: "client_id is required"
      properties:
        error:
          type: string
          description: Error code
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - access_denied
            - server_error
            - invalid_token
            - invalid_dpop_proof
            - expired_token
            - authorization_pending
            - slow_down
            - invalid_redirect_uri
          example: invalid_request
        error_description:
          type: string
          description: Human-readable error description
          example: "client_id is required"
        state:
          type: string
          description: State parameter from the original request (echoed back)
          example: "abc123xyz"